Simplify the backfill migration
